+++ src/Products/urban/UrbanTool.py
@@ -15,8 +15,6 @@ __docformat__ = 'plaintext'
 
 from AccessControl import ClassSecurityInfo
 
-from persistent.mapping import PersistentMapping
-
 from Products.Archetypes.atapi import *
 
 from Products.CMFCore.utils import UniqueObject
@@ -26,11 +24,9 @@ from Products.DataGridField import DataGridField, DataGridWidget
 from Products.DataGridField.Column import Column
 from collective.datagridcolumns.DateColumn import DateColumn
 
-from Products.urban.config import NIS
-from Products.urban.config import URBANMAP_CFG
-from Products.urban.config import VOCABULARY_TYPES
 from Products.urban.config import *
-from Products.urban.interfaces import IUrbanEventType
+from Products.urban.interfaces import IContactFolder
+from Products.urban.interfaces import IUrbanVocabularyTerm
 from Products.urban.interfaces import IGenericLicence
 from Products.urban import UrbanMessage as _
 
@@ -42,15 +38,12 @@ from plone import api
 from plone.memoize.request import cache
 from zope.i18n import translate
 from Products.CMFCore import permissions
-from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.Expression import Expression
 from Products.CMFPlone.i18nl10n import ulocalized_time
 from Products.PageTemplates.Expressions import getEngine
 from Products.DataGridField.DataGridField import FixedRow
 from Products.DataGridField.FixedColumn import FixedColumn
 from Products.urban.utils import getCurrentFolderManager
-from Products.urban.config import GENERATED_DOCUMENT_FORMATS
-from Products.urban.interfaces import IUrbanVocabularyTerm, IContactFolder
 from Products.urban import services
 from datetime import date as _date
 
@@ -279,7 +272,7 @@ class UrbanTool(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
          Return the default text of the field (if it exists)
         """
         if not config:
-            config = getattr(self, self.getUrbanConfig(context).getId())
+            config = getattr(self, self.getLicenceConfig(context).getId())
         for prop in config.getTextDefaultValues():
             if 'fieldname' in prop and prop['fieldname'] == fieldname:
                 return prop['text']
@@ -327,7 +320,7 @@ class UrbanTool(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
         """
         #search in an urbanConfig or in the tool
         if inUrbanConfig:
-            vocPath = "%s/%s/%s" % ('/'.join(self.getPhysicalPath()), self.getUrbanConfig(context).getId(), vocToReturn)
+            vocPath = "%s/%s/%s" % ('/'.join(self.getPhysicalPath()), self.getLicenceConfig(context).getId(), vocToReturn)
         else:
             vocPath = "%s/%s" % ('/'.join(self.getPhysicalPath()), vocToReturn)
         brains = self.portal_catalog(path=vocPath, sort_on=sort_on, portal_type=vocType, review_state=allowedStates)
@@ -473,15 +466,15 @@ class UrbanTool(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
         """
           Return a list of topics to display in the portlet
         """
-        topics = self.getUrbanConfig(context).topics.objectValues('ATTopic')
+        topics = self.getLicenceConfig(context).topics.objectValues('ATTopic')
         res = []
         for topic in topics:
             res.append(topic)
 
         return res
 
-    security.declarePublic('getUrbanConfig')
-    def getUrbanConfig(self, context, urbanConfigId=None):
+    security.declarePublic('getLicenceConfig')
+    def getLicenceConfig(self, context, urbanConfigId=None):
         """
           Return the folder containing the necessary paramaters
         """
@@ -576,28 +569,6 @@ class UrbanTool(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
             return at_obj[fieldRealName]
         else:
             return at_obj.Schema()[fieldRealName]
-
-    security.declarePublic('listEventTypes')
-    def listEventTypes(self, context, urbanConfigId):
-        """
-          Returns the eventTypes of an urbanConfigProxy
-        """
-        urbanConfig = self.getUrbanConfig(context=None, urbanConfigId=urbanConfigId)
-        cat = getToolByName(self, 'portal_catalog')
-        path = '/'.join(urbanConfig.getPhysicalPath())
-        brains = cat(
-            path=path,
-            sort_on='getObjPositionInParent',
-            object_provides=IUrbanEventType.__identifier__,
-            review_state="enabled"
-        )
-        res = []
-        #now evaluate the TAL condition for every brain
-        for brain in brains:
-            event_type = brain.getObject()
-            if event_type.canBeCreatedInLicence(context):
-                res.append(brain)
-        return res
 
     security.declarePublic('decorateHTML')
     def decorateHTML(self, classname, htmlcode):
@@ -783,6 +754,9 @@ class UrbanTool(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
 
     def isContactFolder(self, folder):
         return IContactFolder.providedBy(folder)
+
+    def isLicence(self, context):
+        return IGenericLicence.providedBy(context)
 
     def get_division_name(self, division_code, alternative_name=True):
         mapping = dict([(str(int(l['division'])), l['alternative_name']) for l in self.getDivisionsRenaming()])