+++ src/Products/urban/content/licence/Inspection.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-#
+
 from AccessControl import ClassSecurityInfo
 from Products.Archetypes.atapi import *
 from zope.interface import implements
@@ -370,6 +370,15 @@ def finalize_schema(schema, folderish=False, moveDiscussion=True):
     schema.moveField('description', after='inspection_context')
     schema.moveField('bound_licences', before='workLocations')
     schema.moveField('use_bound_licence_infos', after='bound_licences')
+    schema['parcellings'].widget.label = _('urban_label_parceloutlicences')
+    schema['isInSubdivision'].widget.label = _('urban_label_is_in_parceloutlicences')
+    schema['subdivisionDetails'].widget.label = _('urban_label_parceloutlicences_details')
+    schema['pca'].vocabulary = UrbanVocabulary('sols', vocType="PcaTerm", inUrbanConfig=False)
+    schema['pca'].widget.label = _('urban_label_sol')
+    schema['pcaZone'].vocabulary_factory = 'urban.vocabulary.SOLZones'
+    schema['pcaZone'].widget.label = _('urban_label_solZone')
+    schema['isInPCA'].widget.label = _('urban_label_is_in_sol')
+    schema['pcaDetails'].widget.label = _('urban_label_sol_details')
     return schema