+++ src/Products/urban/indexes.py
@@ -34,7 +34,6 @@ from Products.urban.interfaces import IProprietary
 from Products.urban.interfaces import ITicket
 from Products.urban.interfaces import IUrbanDoc
 from Products.urban.interfaces import IUrbanEvent
-from Products.urban.interfaces import IUrbanEventType
 from Products.urban.schedule.interfaces import ILicenceDeliveryTask
 from Products.urban.utils import get_ws_meetingitem_infos
 
@@ -137,18 +136,6 @@ def genericlicence_lastkeyevent(object):
         event_type = event.getUrbaneventtypes()
         if event_type.getIsKeyEvent() and event.getEventDate().year() >= 1900:
             return "%s,  %s" % (event.getEventDate().strftime("%d/%m/%y"), event_type.Title())
-
-
-# !!!!
-# We use this index to know if an event is schedulable or not.
-# Since it's not used for UrbanEventType, we use this one rather
-# than define a new index
-# !!!!
-@indexer(IUrbanEventType)
-def urbaneventtype_lastkeyevent(object):
-    if object.getDeadLineDelay() > 0:
-        return 'schedulable'
-    return ''
 
 
 @indexer(IGenericLicence)