+++ src/Products/urban/migration/to_DX/migration_utils.py
@@ -16,6 +16,16 @@ def migrate_date(src_obj, dst_obj, src_fieldname, dst_fieldname):
         setattr(dst_obj, dst_fieldname, new_date)
 
 
+def migrate_to_tuple(src_obj, dst_obj, src_fieldname, dst_fieldname):
+    old_value = src_obj.getField(src_fieldname).getRaw(src_obj)
+    new_value = old_value
+    if type(old_value) in [str, unicode]:
+        new_value = (old_value,)
+    elif type(old_value) is list:
+        new_value = tuple(old_value)
+    setattr(dst_obj, dst_fieldname, new_value)
+
+
 def clean_obsolete_portal_type(portal_type_to_remove=None, report='print'):
 
     if not portal_type_to_remove:
@@ -62,3 +72,12 @@ def delete_plone_objects(portal_type_object_to_delete):
     logger.info("Found {} items to be deleted".format(len(items)))
     api.content.delete(objects=items)
     logger.info("***Done ***")
+
+
+def uid_catalog_reindex_objects(objects=[]):
+    """
+    Reindex the given objects the the UID catalog.
+    """
+    uid_catalog = api.portal.get_tool('uid_catalog')
+    for obj in objects:
+        uid_catalog.catalog_object(obj, '/'.join(obj.getPhysicalPath()))