+++ docs/CHANGES.rst
@@ -56,6 +56,12 @@ Changes
 - Migrate ParcellingTerm (Archetype) type to Parcelling (dexterity) type.
   [sdelcourt]
 
+- Add field to define final states closing all the urban events on a licence. [URB-2082]
+  [sdelcourt]
+
+- Refactor key date display to include urban event custom titles. [SUP-13982]
+  [sdelcourt]
+
 
 2.4 (2019-03-25)
 ----------------